refactor(rocketchat): move configuration types into rocketchat package
